# APIs/OpenAPI/openapi-cesium-apis-uKj1o2JVdF76cu6bg5mTDZ.json
{"openapi":"3.0.0","servers":[{"url":"http://example.com","description":"","variables":{}}],"info":{"version":"814638ea-68fe-4ae8-bf19-0c4ba05179d5","title":"cesium-apis","description":"# Introduction\nTest server status\n\n# Overview\nTests status of the server\n\n# Authentication\nNo authentication\n\n# Error Codes\n200","termsOfService":"","contact":{},"license":{"name":""}},"paths":{"/sites/{id}/materials":{"get":{"summary":"Fetch materials","operationId":"Fetchmaterials","parameters":[{"$ref":"#/components/parameters/Content-Type"},{"$ref":"#/components/parameters/id"}],"responses":{"200":{"description":"","headers":{}}}},"post":{"summary":"Create materials","operationId":"Creatematerials","parameters":[{"name":"Content-Type","in":"header","required":true,"style":"simple","schema":{"type":"string","example":"application/json"}},{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"id"}}],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string","example":[{"name":"Gravel","volume_cubic_meter":2.5,"cost_cubic_meter":2,"color":"green","delivery_date":"2021-01-02","construction_site":"{{site_id}}"},{"name":"Dirt","volume_cubic_meter":1.5,"cost_cubic_meter":2,"color":"gray","delivery_date":"2021-01-03","construction_site":"{{site_id}}"}]},"example":"[\n    {\n        \"name\": \"Gravel\",\n        \"volume_cubic_meter\": 2.5,\n        \"cost_cubic_meter\": 2,\n        \"color\": \"green\",\n        \"delivery_date\": \"2021-01-02\",\n        \"construction_site\": \"{{site_id}}\"\n    },\n    {\n        \"name\": \"Dirt\",\n        \"volume_cubic_meter\": 1.5,\n        \"cost_cubic_meter\": 2,\n        \"color\": \"gray\",\n        \"delivery_date\": \"2021-01-03\",\n        \"construction_site\": \"{{site_id}}\"\n    }\n]"}}}},"put":{"summary":"Update materials","operationId":"Updatematerials","parameters":[{"name":"Content-Type","in":"header","required":true,"style":"simple","schema":{"type":"string","example":"application/json"}},{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"id"}}],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string","example":[{"material_id":11,"material":{"name":"Gravel 2","volume_cubic_meter":2.5,"cost_cubic_meter":3.5,"color":"green 2","delivery_date":"2021-01-02","construction_site":"{{site_id}}"}},{"material_id":12,"material":{"name":"Dirt 2","volume_cubic_meter":1.5,"cost_cubic_meter":2,"color":"gray 2","delivery_date":"2021-01-03","construction_site":"{{site_id}}"}}]},"example":"[\n    {\n        \"material_id\": 11,\n        \"material\": {\n            \"name\": \"Gravel 2\",\n            \"volume_cubic_meter\": 2.5,\n            \"cost_cubic_meter\": 3.5,\n            \"color\": \"green 2\",\n            \"delivery_date\": \"2021-01-02\",\n            \"construction_site\": \"{{site_id}}\"\n        }\n    },\n    {\n        \"material_id\": 12,\n        \"material\": {\n            \"name\": \"Dirt 2\",\n            \"volume_cubic_meter\": 1.5,\n            \"cost_cubic_meter\": 2,\n            \"color\": \"gray 2\",\n            \"delivery_date\": \"2021-01-03\",\n            \"construction_site\": \"{{site_id}}\"\n        }\n    }\n]"}}}}},"/sites/{id}/materials/{materialIds}":{"delete":{"summary":"Delete materials","operationId":"Deletematerials","parameters":[{"$ref":"#/components/parameters/materialIds"},{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"id"}}],"responses":{"200":{"description":"","headers":{}}}}}},"components":{"parameters":{"Content-Type":{"name":"Content-Type","in":"header","required":true,"style":"simple","schema":{"type":"string","example":"application/json"}},"id":{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"id"}},"materialIds":{"name":"materialIds","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"[11]"}}}},"security":[],"tags":[],"externalDocs":{"url":"","description":""},"warnings":[]}